+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0
 
 * Saved byte data as bytes and not as base64encoded to be more space efficient
+* Removed retrieveAll since don't know if it is string data or byte data
 
 ## 0.2.0